_src/requirements.txt
boto3==1.34.144
python-dotenv==1.0.1
tqdm==4.66.1
nltk>=3.8.1
python-frontmatter>=1.0.0
pydantic>=2.0.0